Maze window moves through all screens
Also the code is ready to share the textures between windows (but it's
not ready to have all monitors with the maze switching between windows -
one of the monitors remains with a black window while the other 2 have
the maze).

One CPU core is at 100% though, because of PeekMessage instead of
GetMessage. That must be fixed.

Also the monitors resolutions I think must be the same, because I
reduced one of them and the maze got too big to fit on the screen.